attractions_django/requirements.txt
djangorestframework==3.16.0